+++ src/tweetz.core/ViewModels/HomeTimelineControlViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using tweetz.core.Infrastructure;
 using twitter.core.Models;
@@ -19,7 +20,33 @@ namespace tweetz.core.ViewModels
 
         protected override async Task<IEnumerable<TwitterStatus>> GetTimeline()
         {
-            return await TwitterService.GetHomeTimeline();
+            var mentions = await GetMentions();
+            var statuses = await TwitterService.GetHomeTimeline();
+            return statuses.Concat(mentions);
+        }
+
+        // Twitter limits getting mentions to 100,000 per day per Application.
+        // Application in this case means all running Tweetz clients. Once
+        // an hour allows everybody get mentions albiet not in a timely manner.
+        private const int mentionsInterval = 60;
+
+        private int mentionsCounter = mentionsInterval;
+
+        private async Task<IEnumerable<TwitterStatus>> GetMentions()
+        {
+            IEnumerable<TwitterStatus> mentions = new TwitterStatus[0];
+
+            if (mentionsCounter >= mentionsInterval)
+            {
+                mentionsCounter = 0;
+                mentions = await TwitterService.GetMentionsTimeline();
+            }
+            else
+            {
+                mentionsCounter += 1;
+            }
+
+            return mentions;
         }
     }
 }